Added sketch file
There is room for improvement I'm sure.  Such as using a 20x4 LCD would cut used dynamic memory space and make it alot more readable at-a-glance.  There then should be room for a rainfall measurement along with wind-speed and direction.
Not sure you could then use a network/data logger to that, but maybe.